coin_market_scrap/requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
Django
django-celery-beat==2.6.0
django-rest-framework==0.1.0
django-timezone-field==6.1.0
djangorestframework==3.15.1
exceptiongroup==1.2.1
h11==0.14.0
idna==3.7
kombu==5.3.7
outcome==1.3.0.post0
prompt-toolkit==3.0.43
PySocks==1.7.1
python-crontab==3.0.0
python-dateutil==2.9.0.post0
requests==2.31.0
selenium==4.19.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.0
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
wsproto==1.2.0
